[CDSR-1726] move single only controller into the overpaymentssingle package
