+++ app/uk/gov/hmrc/cdsreimbursementclaimfrontend/views/claims/enter_single_claim.scala.html
@@ -22,7 +22,8 @@
 @import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.actions.RequestWithSessionData
 @import uk.gov.hmrc.cdsreimbursementclaimfrontend.models.ClaimedReimbursement
 @import uk.gov.hmrc.cdsreimbursementclaimfrontend.models.BigDecimalOps
-@import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.claims.EnterSingleClaimController.ClaimAmount
+@import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.overpaymentssingle.EnterSingleClaimController.ClaimAmount
+@import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.overpaymentssingle.{routes=>overpaymentsSingleRoutes}
 @import uk.gov.hmrc.govukfrontend.views.viewmodels.content.Text
 @import uk.gov.hmrc.govukfrontend.views.viewmodels.input.PrefixOrSuffix
 @import java.util.UUID
@@ -54,7 +55,7 @@
 
     @paragraph(Html(messages(s"$key.inset-text")), Some("govuk-inset-text govuk-!-margin-bottom-8"))
 
-    @formWithCSRF(routes.EnterSingleClaimController.enterClaimSubmit(id), 'novalidate -> "novalidate") {
+    @formWithCSRF(overpaymentsSingleRoutes.EnterSingleClaimController.enterClaimSubmit(id), 'novalidate -> "novalidate") {
 
         @paragraph(Html(messages(s"$key.paid-amount-label", reimbursement.paidAmount.toPoundSterlingString)), Some("govuk-body-l govuk-!-font-weight-bold"))
